requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.0
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
whitenoise==6.4.0
